Modified project structure by applying ES module convention (#1)
