+++ src/index.html
@@ -12,7 +12,7 @@
       href="https://js.arcgis.com/4.17/esri/themes/light/main.css"
     />
     <script src="https://js.arcgis.com/4.17/"></script>
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
     <script src="js/nav.js"></script>
     <link rel="stylesheet" type="text/css" href="css/main.css" />
   </head>
