+++ src/js/config.js
@@ -1,8 +1,25 @@
-const kelpProductivityLayerUrl = "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/ussw11999_maxcanopy/FeatureServer/0";
-const shippingLanesLayerUrl = "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/ShippingLanes_SCA/FeatureServer/1";
-const dangerZonesAndRestrictedAreasLayerUrl = "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/DangerZonesAndRestrictedAreas_SCA/FeatureServer/0";
-const mpaInventoryLayerUrl = "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/MPAInventory_SCA/FeatureServer/0";
+const kelpProductivityLayerUrl =
+  "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/ussw11999_maxcanopy/FeatureServer/0";
+const shippingLanesLayerUrl =
+  "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/ShippingLanes_SCA/FeatureServer/1";
+const dangerZonesAndRestrictedAreasLayerUrl =
+  "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/DangerZonesAndRestrictedAreas_SCA/FeatureServer/0";
+const mpaInventoryLayerUrl =
+  "https://services7.arcgis.com/4c8njmg1eMIbzYXM/arcgis/rest/services/MPAInventory_SCA/FeatureServer/0";
+const principalPortsLayerUrl =
+  "https://services7.arcgis.com/4c8njmg1eMIbzYXM/ArcGIS/rest/services/PrincipalPorts_SCA/FeatureServer/0";
+const federalAndStateWatersLayerUrl =
+  "https://services7.arcgis.com/4c8njmg1eMIbzYXM/ArcGIS/rest/services/FederalAndStateWaters/FeatureServer/0";
+
+const locatorTaskUrl =
+  "https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer";
 
 export {
-    kelpProductivityLayerUrl, shippingLanesLayerUrl, dangerZonesAndRestrictedAreasLayerUrl, mpaInventoryLayerUrl
-}
+  kelpProductivityLayerUrl,
+  shippingLanesLayerUrl,
+  dangerZonesAndRestrictedAreasLayerUrl,
+  mpaInventoryLayerUrl,
+  principalPortsLayerUrl,
+  federalAndStateWatersLayerUrl,
+  locatorTaskUrl,
+};
